backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "flask>=2.0",
    "flask-cors>=3.0",
    "flask-sqlalchemy>=2.5",
    "groq>=0.30.0",
    "langchain>=0.0.200",
    "langchain-community>=0.0.50",
    "langchain-experimental>=0.0.10",
    "langchain-groq>=0.0.10",
    "marshmallow>=3.0",
    "pandas>=1.3",
    "python-dotenv>=0.19",
    "sqlalchemy>=1.4",
    "tabulate>=0.9.0",
]
